rename: user 패키지 구조 평탄화 및 파일 이동 - 다른 패키지들과 동일한 구조로 변경
