+++ src/main/java/com/capstone/domain/user/UserRepository.java
@@ -1,6 +1,5 @@
-package com.capstone.domain.user.repository;
+package com.capstone.domain.user;
 
-import com.capstone.domain.user.entity.User;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
