+++ src/main/java/com/capstone/domain/workspaceInvitation/WorkspaceInvitationRepository.java
@@ -1,6 +1,6 @@
 package com.capstone.domain.workspaceInvitation;
 
-import com.capstone.domain.user.entity.User;
+import com.capstone.domain.user.User;
 import com.capstone.domain.workspace.Workspace;
 import java.util.List;
 import java.util.Optional;
