+++ src/test/java/com/capstone/integration/ChatMessageIntegrationTest.java
@@ -2,7 +2,7 @@ package com.capstone.integration;
 
 import com.capstone.domain.chat.ChatMessage;
 import com.capstone.domain.chat.ChatMessageRepository;
-import com.capstone.domain.user.entity.User;
+import com.capstone.domain.user.User;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
