+++ src/test/java/com/capstone/service/CanvasServiceTest.java
@@ -5,8 +5,8 @@ import com.capstone.domain.canvas.CanvasRepository;
 import com.capstone.domain.canvas.CanvasRequest;
 import com.capstone.domain.canvas.CanvasResponse;
 import com.capstone.domain.canvas.CanvasService;
-import com.capstone.domain.user.entity.User;
-import com.capstone.domain.user.repository.UserRepository;
+import com.capstone.domain.user.User;
+import com.capstone.domain.user.UserRepository;
 import com.capstone.domain.workspace.Workspace;
 import com.capstone.domain.workspace.WorkspaceRepository;
 import com.capstone.domain.workspaceUser.WorkspaceUser;
