+++ src/test/java/com/capstone/service/ChatMessageServiceTest.java
@@ -3,8 +3,8 @@ package com.capstone.service;
 import com.capstone.domain.chat.ChatMessage;
 import com.capstone.domain.chat.ChatMessageRepository;
 import com.capstone.domain.chat.ChatMessageService;
-import com.capstone.domain.user.entity.User;
-import com.capstone.domain.user.repository.UserRepository;
+import com.capstone.domain.user.User;
+import com.capstone.domain.user.UserRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
